move Users functionality to separate file
